+++ src/main/java/org/mclods/factory_method_pattern/factory_pattern.md
